requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
pyyaml==6.0.1
